+++ packages/dumb/src/models/safety-basal-item.model.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2025, Diabeloop
+ * Copyright (c) 2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,31 +25,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import * as auth0Mock from '@auth0/auth0-react'
-import { getAccessTokenWithPopupMock, logoutMock } from '../../mock/auth0.hook.mock'
-import { renderPage } from '../../utils/render'
-import { waitFor } from '@testing-library/react'
-import { AppRoute } from '../../../../models/enums/routes.enum'
-import { testVerifyEmail } from '../../use-cases/email-verification'
-import { Auth0Error } from '../../../../lib/auth/models/enums/auth0-error.enum'
-
-describe('Verify email page', () => {
-  it('should display a description of the email verification process with options', async () => {
-    (auth0Mock.useAuth0 as jest.Mock).mockReturnValue({
-      isAuthenticated: false,
-      isLoading: false,
-      user: null,
-      getAccessTokenWithPopup: getAccessTokenWithPopupMock,
-      logout: logoutMock,
-      getAccessTokenSilently: jest.fn().mockRejectedValue({ error_description: Auth0Error.EmailNotVerified })
-    })
-    window.open = jest.fn()
-
-    const router = renderPage(AppRoute.VerifyEmail)
-    await waitFor(() => {
-      expect(router.state.location.pathname).toEqual(AppRoute.VerifyEmail)
-    })
-
-    await testVerifyEmail()
-  })
-})
+export interface SafetyBasalItem {
+  rate: string,
+  startTime: string,
+  endTime: string
+}