+++ packages/yourloops/utils/error.util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2025, Diabeloop
+ * Copyright (c) 2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,31 +25,30 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import * as auth0Mock from '@auth0/auth0-react'
-import { getAccessTokenWithPopupMock, logoutMock } from '../../mock/auth0.hook.mock'
-import { renderPage } from '../../utils/render'
-import { waitFor } from '@testing-library/react'
-import { AppRoute } from '../../../../models/enums/routes.enum'
-import { testVerifyEmail } from '../../use-cases/email-verification'
-import { Auth0Error } from '../../../../lib/auth/models/enums/auth0-error.enum'
+import { browserName, browserVersion } from 'react-device-detect'
+import moment from 'moment-timezone'
+import { v4 as uuidv4 } from 'uuid'
+import ErrorApi, { ErrorPayload } from '../lib/error/error.api'
+import metrics from '../lib/metrics'
 
-describe('Verify email page', () => {
-  it('should display a description of the email verification process with options', async () => {
-    (auth0Mock.useAuth0 as jest.Mock).mockReturnValue({
-      isAuthenticated: false,
-      isLoading: false,
-      user: null,
-      getAccessTokenWithPopup: getAccessTokenWithPopupMock,
-      logout: logoutMock,
-      getAccessTokenSilently: jest.fn().mockRejectedValue({ error_description: Auth0Error.EmailNotVerified })
-    })
-    window.open = jest.fn()
+export const logError = (errorMessage: string, context: string): void => {
+  try {
+    metrics.send('error', context, errorMessage)
+    const error = getErrorObject(errorMessage)
 
-    const router = renderPage(AppRoute.VerifyEmail)
-    await waitFor(() => {
-      expect(router.state.location.pathname).toEqual(AppRoute.VerifyEmail)
-    })
+    ErrorApi.sendError(error)
+  } catch (err) {
+    console.error(err)
+  }
+}
 
-    await testVerifyEmail()
-  })
-})
+const getErrorObject = (errorMessage: string): ErrorPayload => {
+  return {
+    browserName,
+    browserVersion,
+    date: moment(new Date()).format('DD/MM/YYYY'),
+    err: errorMessage,
+    errorId: uuidv4(),
+    path: location.pathname
+  }
+}